Cuba 7 rewriting

+++ modules/web/src/com/joker/jokerapp/web/dialogs/ActualSeatsDialog.java
@@ -50,7 +50,7 @@ public class ActualSeatsDialog extends AbstractWindow {
 
         if (handler != null) {
 
-            int seatsNum = seatsTextField.getValue();
+            int seatsNum = (int) seatsTextField.getValue();
             handler.onClose(seatsNum);
 
         }
